Refactor property page and add property details
