+++ client/src/app/(dashboard)/profile/[id]/landlords/properties/[propertyId]/page.tsx
@@ -1,11 +1,12 @@
-import PropertyForm from "@/components/shared/properties/propertyForm";
+import { Button } from "@/components/ui/button";
 import type { Property } from "@/types";
 import axios from "axios";
+import Link from "next/link";
 
-const UpdateProperty = async ({
+const PropertyDetails = async ({
   params,
 }: {
-  params: { propertyId: string };
+  params: { id: string; propertyId: string };
 }) => {
   const property = await axios.get<Property>(
     `${process.env.NEXT_PUBLIC_API_URL}/properties/${params.propertyId}`,
@@ -16,14 +17,22 @@ const UpdateProperty = async ({
   return (
     <div className="w-full">
       <h1 className="mb-2 mt-4 text-3xl font-bold text-indigo-950">
-        Update Property
+        Property Details
       </h1>
       <p className="mb-4 max-w-xl text-indigo-950">
-        Update your property details to match your needs.
+        This page contains all the details of your property {property.data.name}
+        .
       </p>
-      <PropertyForm action="Update" propertyData={property.data} />
+      <p className="mb-4 max-w-xl text-indigo-950">
+        {JSON.stringify(property.data)}
+      </p>
+      <Link
+        href={`/profile/${params.id}/landlords/properties/${params.propertyId}/update`}
+      >
+        <Button variant="outline">Update Property</Button>
+      </Link>
     </div>
   );
 };
 
-export default UpdateProperty;
+export default PropertyDetails;
